Refactor ElementResultsWriter: * Stub out TimeSeriesPlotWriter
TimeSeriesPlotWriter:
* Fix bug where PrecipAndOutflow plots are truncated
* Simplify logic
